.history/requirements_20250113135220.txt
keyboard==0.13.5
pywin32==306
python-dotenv==1.0.0
pystray
Pillow
